Refactor libraries and multline source/destination
Libraries are refactored to seperate files/modules and the multiline
option has been added to the source/destination builder method.

Refactor method names

Update method calls

Initial multiline testing with source library/template/resource.

[ci skip]
