+++ resources/destination.rb
@@ -48,7 +48,7 @@ action :create do
       description: new_resource.description.nil? ? new_resource.name : new_resource.description,
       destination: destination
     )
-    helpers(SyslogNg::ConfigHelpers)
+    helpers(SyslogNg::DestinationHelpers)
     action :create
   end
 end
