+++ templates/default/syslog-ng/destination.conf.erb
@@ -8,7 +8,7 @@
 <% @destination.each do |name, config| -%>
 destination <%= name -%> {
     <% config.each do |driver, parameters| -%>
-    <%= config_srcdst_driver_map(driver, parameters) %>
+    <%= destination_builder(driver: driver, parameters: parameters) %>
     <% end -%>
 };
 <% end -%>
